fix(sonar): S3776 AISection.tsx (cx 181)
Extract provider panels (dome/copilot/claude-oauth/openai-codex), chat
tab body, save bar, controller hook, and load/save helpers so the AISection
default export stays under Sonar cognitive complexity threshold. Behavior
preserved; small unit tests cover pure helpers.

Sonar key: 31b87bab-0cdc-49b5-a383-bd45929fc946
